Replace relative document links with absolute ones
There is an issue in Lacuna's dashboard that causes relative document links to open as https://lacunabot.com/guilds/GUILD_ID/settings/docs/FILENAME.md. This commit fixes such behaviour, so that document links should now correctly point to GitHub. The only downside to this fix is that links in forked repos will always lead to documents in the parent repo, this needs to be changed manually for each fork.